pyproject.toml
[tool.poetry]
name = "ticktok"
version = "0.1.0"
description = "Official Python client for Ticktok.io"
authors = ["elis"]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pytest = "^6.1.1"
invoke = "^1.4.1"
PyHamcrest = "^2.0.2"
requests = "^2.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

